+++ src/commands.ts
@@ -1,17 +1,17 @@
 import * as vscode from 'vscode';
-import * as selection from './selection';
+import * as cursor from './cursor';
 import * as folding from './folding';
 import * as events from './events';
 
 export function registerCommands(context: vscode.ExtensionContext) {
     console.log('Registering commands');
     context.subscriptions.push(vscode.commands.registerCommand(
-        'bmx-local.selectLineUp', (args: {}) => {
-            selection.selectLineUp(args);
+        'bmx-local.cursorUp', (args: {}) => {
+            cursor.cursorUp(args);
         }));
     context.subscriptions.push(vscode.commands.registerCommand(
-        'bmx-local.selectLineDown', (args: {}) => {
-            selection.selectLineDown(args);
+        'bmx-local.cursorDown', (args: {}) => {
+            cursor.cursorDown(args);
         }));
     context.subscriptions.push(vscode.commands.registerCommand(
         'bmx-local.multiStateFold', async(args: {}) => {