+++ src/cursor.ts
@@ -19,7 +19,7 @@ function getActiveLineRange(editor: vscode.TextEditor) {
     let activeLineEndChar = endChar(editor, activeLine);
     return new vscode.Range(
         new vscode.Position(activeLine, 0),
-        new vscode.Position(activeLine, activeLineEndChar));    
+        new vscode.Position(activeLine, activeLineEndChar));
 }
 
 function adjustedPositionAt(editor: vscode.TextEditor, line: number, location: Location) {
@@ -60,7 +60,7 @@ function isPartialSelection(editor: vscode.TextEditor, dir: number) {
 function selectLine(editor: vscode.TextEditor, dir: number, amount: number) {
     let anchorLine = editor.selection.anchor.line;
     let activeLine = editor.selection.active.line;
-    
+
     // if current line is not fully selected, select one less line in total
     let adjustedAmount = amount;
     if (isPartialSelection(editor, dir)) {
@@ -79,11 +79,11 @@ function selectLine(editor: vscode.TextEditor, dir: number, amount: number) {
     if (isNewSelectionReversed) {
         newSelection = new vscode.Selection(
             adjustedPositionAt(editor, anchorLine, Location.End),
-            adjustedPositionAt(editor, activeLine + dir*adjustedAmount, Location.Start));
+            adjustedPositionAt(editor, activeLine + dir * adjustedAmount, Location.Start));
     } else {
         newSelection = new vscode.Selection(
             adjustedPositionAt(editor, anchorLine, Location.Start),
-            adjustedPositionAt(editor, activeLine + dir*adjustedAmount, Location.End));
+            adjustedPositionAt(editor, activeLine + dir * adjustedAmount, Location.End));
     }
 
     if (editor.selection.isEmpty ||
@@ -101,39 +101,78 @@ function selectLine(editor: vscode.TextEditor, dir: number, amount: number) {
     editor.revealRange(getActiveLineRange(editor));
 }
 
-export function selectLineUp(args: {}) {
-    let editor = vscode.window.activeTextEditor;
-    if (editor === undefined) {
-        return;
+function moveCursor(editor: vscode.TextEditor, dir: number, amount: number) {
+    let newActiveLine = editor.selection.active.line + dir * amount;
+    let newActiveChar = editor.selection.active.character;
+
+    // adjust values if line is out of boundaries
+    let lastLine = editor.document.lineCount - 1;
+    if (newActiveLine < 0) {
+        newActiveLine = 0;
+        newActiveChar = 0;
+    } else if (newActiveLine > lastLine) {
+        newActiveLine = lastLine;
+        newActiveChar = endChar(editor, lastLine);
     }
 
-    let amount = 1;
-    if (args !== undefined &&
-        'amount' in args &&
-        typeof(args['amount']) === 'number' &&
-        args['amount'] > 0
-    ) {
-        amount = args['amount'];
-    }
-
-    selectLine(editor, -1, amount);
+    editor.selection = new vscode.Selection(
+        new vscode.Position(newActiveLine, newActiveChar),
+        new vscode.Position(newActiveLine, newActiveChar)
+    );
+    editor.revealRange(getActiveLineRange(editor));
 }
 
-export function selectLineDown(args: {}) {
+function cursor(args: {}, dir: number) {
     let editor = vscode.window.activeTextEditor;
     if (editor === undefined) {
         return;
     }
 
+    // Parse args
     let amount = 1;
-    if (args !== undefined &&
-        'amount' in args &&
-        typeof(args['amount']) === 'number' &&
-        args['amount'] > 0
-    ) {
-        amount = args['amount'];
+    let select = false
+    if (args !== undefined) {
+        // Calculate requested amount
+        if ('amount' in args && typeof (args['amount']) === 'number'
+            && args['amount'] > 0
+        ) {
+            amount = args['amount'];
+        } else if ('amountPercent' in args && typeof (args['amountPercent']) === 'number' &&
+            args['amountPercent'] > 0 && args['amountPercent'] <= 100
+        ) {
+            amount = Math.round(editor.document.lineCount * args['amountPercent'] / 100)
+        }
+
+        // Clamp amount to appropriate range
+        if ('amountMin' in args && typeof (args['amountMin']) === 'number'
+            && args['amountMin'] > 0
+        ) {
+            amount = (amount >= args['amountMin']) ? amount : args['amountMin'];
+        }
+        if ('amountMax' in args && typeof (args['amountMax']) === 'number'
+            && args['amountMax'] > 0
+        ) {
+            amount = (amount <= args['amountMax']) ? amount : args['amountMax'];
+        }
+
+        // Should select?
+        if ('select' in args && typeof (args['select']) === 'boolean') {
+            select = args['select'];
+        }
     }
 
-    selectLine(editor, 1, amount);
+    // Delegate to appropriate function
+    if (select) {
+        selectLine(editor, dir, amount);
+    } else {
+        moveCursor(editor, dir, amount);
+    }
+}
 
+export function cursorUp(args: {}) {
+    cursor(args, -1)
+}
+
+export function cursorDown(args: {}) {
+    cursor(args, 1)
 }